c: loops

+++ src/sem-1/UE23CS151B/unit_1/calc.c
@@ -24,7 +24,7 @@ int main()
             printf("%d", a*b);
             break;
     case '/': 
-            printf("%d", a/b);
+            printf("%f", (float)a/(float)b);
             break;
 
     default: printf("%s", "error\n");
